requirements.txt
streamlit
pandas
plotly
prophet
scikit-learn
numpy
matplotlib
google-generativeai
